+++ core/src/main/java/at/campus02/swd/game/game/DefenceStrategy.java
@@ -1,4 +1,4 @@
-package at.campus02.swd.game.board;
+package at.campus02.swd.game.game;
 
 @FunctionalInterface
 public interface DefenceStrategy {
