+++ core/src/main/java/at/campus02/swd/game/threats/AmplifiedEdgeDamage.java
@@ -1,7 +1,7 @@
 package at.campus02.swd.game.threats;
 
-import at.campus02.swd.game.board.BoardView;
-import at.campus02.swd.game.board.ThreatStrategy;
+import at.campus02.swd.game.game.BoardView;
+import at.campus02.swd.game.game.ThreatStrategy;
 import at.campus02.swd.game.util.Position;
 
 import java.util.*;
